+++ python/pycylon/pycylon/net/oob_context.pxd
@@ -12,11 +12,11 @@
  # limitations under the License.
  ##
 
-IF CYTHON_UCX & CYTHON_UCC:
+IF CYTHON_UCX :
 
     from pycylon.common.status cimport CStatus
 
-    cdef extern from "../../../../cpp/src/cylon/net/ucx/ucx_ucc_oob_contexts.hpp" namespace "cylon::net":
+    cdef extern from "../../../../cpp/src/cylon/net/ucx/ucx_ucc_oob_context.hpp" namespace "cylon::net":
         cdef cppclass CUCXOOBContext "cylon::net::UCXOOBContext":
             CStatus getWorldSizeAndRank(int &world_size, int &rank)
     cdef class UCXOOBContext:
